refactor(agents): service-shaped public agent profile reconcile
- interface + factory instead of a bare function
- compose returns reconcileService, built once
